Update RuleList 'rules' to support expected types
The scala API allows the RuleList 'rules' to be either a RuleList or a PriorityRule, with recursive nesting possible on the RuleLists.
This patch adds (de)serialization support of the RuleList objects
written by the scala backend. The previous java code allowed only the
PriorityRule as the child of a RuleList.

For example, the below JSON is a RuleList with rules being PriorityRule
and a nested RuleList:
client code:
```
{
  "condition": "AND",
  "rules": [
    { // priority rule
      "value": "t.u",
      "type": "string",
      "operator": "is_empty"
    },
    { // rule list
      "condition": "OR",
      "rules": [
        {
          "value": "a.b",
          "type": "string",
          "operator": "is_not_empty"
        }
      ]
    }
  ]
}
```