Support (de)serialization of the RuleList
The maproulette backend expects a string escaped json for all RuleLists in Challenge POSTs, and Challenge GETs serialize the RuleList as an object.
This patch allows the RuleList class to write as a json string and be read back from either a json string or an object.

Also, this enables unit tests to run with the 'gradle build'.