Refactor dependency versions to use top-level ext block (#105)
This commit updates the build.gradle file to define the dependency versions in the top-level `ext` block, using new variable names that follow the format `dep<DependencyName>Version`. The old `versions` block is removed. The `dependencies` block is updated to use the new variable names where appropriate.

The change makes it easier for intellij to scan dependencies and simplifies human updates to the deps.